refactor: rename Footer file
