+++ src/app/App.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Header } from 'common/components/Header';
-import { Footer } from 'common/components/Footer';
+import { Footer } from 'common/components/BakeFooter';
 import { GameSprintSelectLvl } from 'common/components/GameSprintSelectLvl';
 import { Statistic } from 'common/components/Statistic';
 import { HomePage } from 'app/pages/HomePage';
